broke down mealPlanner.tsx
