+++ frontend/src/components/MakeMealPlan.tsx
@@ -1,0 +1,70 @@
+const DAYS = [
+  "monday",
+  "tuesday",
+  "wednesday",
+  "thursday",
+  "friday",
+  "saturday",
+  "sunday",
+];
+
+type Props = {
+  mealCounts: any;
+  setMealCounts: any;
+  recipePreferences: any;
+  setRecipePreferences: any;
+  handleGenerate: any;
+  savedRecipes: { id: string; title: string }[];
+};
+
+export default function MakeMealPlan({
+  mealCounts,
+  setMealCounts,
+  recipePreferences,
+  setRecipePreferences,
+  handleGenerate,
+}: Props) {
+  return (
+    <div>
+      <h3>meals per day</h3>
+      {DAYS.map((day) => (
+        <div key={day}>
+          {day}:
+          <input
+            type="number"
+            value={mealCounts[day]}
+            min={0}
+            max={6}
+            onChange={(e) =>
+              setMealCounts({ ...mealCounts, [day]: +e.target.value })
+            }
+          />
+        </div>
+      ))}
+
+      <h3>recipe preferences</h3>
+      {recipePreferences.length === 0 ? (
+        <p>no saved recipes</p>
+      ) : (
+        recipePreferences.map((r: any, i: number) => (
+          <div key={r.spoonacularId}>
+            {r.title}: eat
+            <input
+              type="number"
+              value={r.servings}
+              min={0}
+              max={10}
+              onChange={(e) => {
+                const updated = [...recipePreferences];
+                updated[i].servings = +e.target.value;
+                setRecipePreferences(updated);
+              }}
+            />
+            times
+          </div>
+        ))
+      )}
+      <button onClick={handleGenerate}>generate meal plan</button>
+    </div>
+  );
+}
